backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.100.0
pydantic[email]==2.1.1
Flask==2.3.2
Flask-CORS==4.0.0
Werkzeug==2.3.6

# Database and ORM
sqlalchemy[asyncio]==2.0.19
asyncpg==0.28.0
supabase==2.0.3
psycopg2-binary==2.9.6

# HTTP and API
requests==2.31.0
urllib3==2.0.4

# Authentication and Security
python-dotenv==1.0.0
PyJWT==2.8.0

# Data Processing
python-dateutil==2.8.2
pytz==2023.3

# Web Server
gunicorn==21.2.0
gevent==23.9.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality
black==23.7.0
flake8==6.0.0
mypy==1.4.1

uvicorn>=0.38.0
# Async Support
aiohttp>=3.7.5
asyncio-contextmanager>=1.0.0
